proper handling of delete operation:
* adding ObjectStatus
* adding virtual delete
* ability to update object

+++ src/main/scala/com/loyalty/testing/s3/service/ObjectService.scala
@@ -8,7 +8,7 @@ import akka.stream.IOResult
 import akka.stream.scaladsl.Source
 import akka.util.ByteString
 import com.loyalty.testing.s3._
-import com.loyalty.testing.s3.repositories.model.{Bucket, ObjectKey, UploadInfo}
+import com.loyalty.testing.s3.repositories.model.{Bucket, ObjectKey, ObjectStatus, UploadInfo}
 import com.loyalty.testing.s3.repositories.{NitriteDatabase, ObjectIO}
 import com.loyalty.testing.s3.request.PartInfo
 import org.slf4j.LoggerFactory
@@ -34,24 +34,34 @@ class ObjectService(objectIO: ObjectIO, database: NitriteDatabase) {
                 (implicit ec: ExecutionContext): Future[ObjectKey] =
     for {
       sourceObjectKey <- objectIO.saveObject(bucket, key, keyId, versionIndex, contentSource)
-      finalObjectKey <- database.createObject(sourceObjectKey)
+      finalObjectKey <- createOrUpdateObject(sourceObjectKey)
     } yield finalObjectKey
+
+  def createOrUpdateObject(objectKey: ObjectKey): Future[ObjectKey] = database.createOrUpdateObject(objectKey)
 
   def getObject(objectKey: ObjectKey,
                 maybeRange: Option[ByteRange] = None): (ObjectKey, Source[ByteString, Future[IOResult]]) =
     objectIO.getObject(objectKey, maybeRange)
 
-  def deleteObject(maybeObjectKey: Option[ObjectKey])
-                  (implicit ec: ExecutionContext): Future[ObjectKey] =
-    maybeObjectKey match {
-      case None => Future.failed(NoSuchKeyException)
-      case Some(objectKey) =>
-        val permanentDelete = objectKey.deleteMarker.isDefined
-        for {
-          _ <- database.deleteObject(objectKey.id, Some(objectKey.versionId), permanentDelete)
-          _ <- deleteFile(objectKey)
-        } yield objectKey.copy(deleteMarker = Some(permanentDelete))
+  def virtualDeleteObject(objectKey: ObjectKey)
+                         (implicit ec: ExecutionContext): Future[ObjectKey] = {
+    val maybeUploadId = objectKey.uploadId
+    if (objectKey.isDeleted) Future.failed(NoSuchKeyException)
+    else {
+      val (status, deleteMarker) = objectKey.status match {
+        case ObjectStatus.Active => (ObjectStatus.Deleted, None)
+        case ObjectStatus.DeleteMarker => (ObjectStatus.DeleteMarkerDeleted, Some(true))
+        case s => throw new RuntimeException(s"Invalid object status: ${objectKey.bucketName}/${objectKey.key}/$s")
+      }
+      val updatedObjectKey = objectKey.copy(eTag = None, contentMd5 = None, contentLength = 0L, status = status,
+        uploadId = None, objectPath = None, deleteMarker = deleteMarker)
+      for {
+        ok <- database.createOrUpdateObject(updatedObjectKey)
+        _ <- deleteFile(objectKey)
+        _ <- database.deleteUpload(maybeUploadId)
+      } yield ok
     }
+  }
 
   def createUpload(uploadInfo: UploadInfo)
                   (implicit ec: ExecutionContext): Future[Done] =
@@ -87,7 +97,7 @@ class ObjectService(objectIO: ObjectIO, database: NitriteDatabase) {
       } else {
         for {
           objectKey <- objectIO.mergeFiles(uploadInfo, parts)
-          savedObjectKey <- database.createObject(objectKey)
+          savedObjectKey <- database.createOrUpdateObject(objectKey)
           _ <- objectIO.moveParts(savedObjectKey, uploadInfo)
           _ <- database.moveParts(savedObjectKey)
         } yield savedObjectKey
@@ -96,21 +106,18 @@ class ObjectService(objectIO: ObjectIO, database: NitriteDatabase) {
 
   def deleteUpload(uploadId: String, partNumber: Int): Future[Done] = database.deleteUpload(uploadId, partNumber)
 
-  private def deleteFile(objectKey: ObjectKey): Future[Done] = {
-    val deleteFile = objectKey.deleteMarker.getOrElse(false)
-    if (deleteFile) {
-      Try(objectIO.delete(objectKey)) match {
-        case Failure(ex) =>
-          // should we penalize client if we are unable to delete file, let's consume exception,
-          // log and move on
-          log.warn(
-            s"""unable to delete file, bucket_name=${objectKey.bucketName}, key=${objectKey.key},
-               | version_id=${objectKey.actualVersionId}""".stripMargin.replaceNewLine, ex)
-          Future.successful(Done)
-        case Success(_) => Future.successful(Done)
-      }
-    } else Future.successful(Done)
-  }
+
+  private def deleteFile(objectKey: ObjectKey): Future[Done] =
+    Try(objectIO.delete(objectKey)) match {
+      case Failure(ex) =>
+        // should we penalize client if we are unable to delete file, let's consume exception,
+        // log and move on
+        log.warn(
+          s"""unable to delete file, bucket_name=${objectKey.bucketName}, key=${objectKey.key},
+             | version_id=${objectKey.actualVersionId}""".stripMargin.replaceNewLine, ex)
+        Future.successful(Done)
+      case Success(_) => Future.successful(Done)
+    }
 }
 
 object ObjectService {
